+++ crates/rtlola-hir/src/modes.rs
@@ -10,7 +10,7 @@ use std::time::Duration;
 use rtlola_reporting::RtLolaError;
 
 use self::dependencies::{
-    DependencyGraph, Origin, Streamdependencies, Transitivedependencies, Windowdependencies,
+    DependencyGraph, Origin, StreamDependencies, TransitiveDependencies, WindowDependencies,
 };
 use self::types::HirType;
 use crate::config::FrontendConfig;
@@ -199,12 +199,12 @@ impl Hir<TypedMode> {
 /// Represents the results of the dependency analysis
 #[derive(Debug, Clone)]
 pub struct DepAna {
-    direct_accesses: Streamdependencies,
-    transitive_accesses: Transitivedependencies,
-    direct_accessed_by: Streamdependencies,
-    transitive_accessed_by: Transitivedependencies,
-    aggregated_by: Windowdependencies,
-    aggregates: Windowdependencies,
+    direct_accesses: StreamDependencies,
+    transitive_accesses: TransitiveDependencies,
+    direct_accessed_by: StreamDependencies,
+    transitive_accessed_by: TransitiveDependencies,
+    aggregated_by: WindowDependencies,
+    aggregates: WindowDependencies,
     graph: DependencyGraph,
 }
 
@@ -272,13 +272,13 @@ pub trait DepAnaTrait {
     ///
     /// The function returns all windows that aggregate `who` and the stream that uses the window.
     /// The result contains only the windows that are direct.
-    fn aggregated_by(&self, who: SRef) -> Vec<(SRef, WRef)>; // (non-transitive)
+    fn aggregated_by(&self, who: SRef) -> Vec<(SRef, Origin, WRef)>; // (non-transitive)
 
     /// Returns all windows that are used in `who` and the corresponding stream that is aggregated
     ///
     /// The function returns all windows that are used in `who` and the corresponding stream that is aggregated.
     /// The result contains only the windows that are direct.
-    fn aggregates(&self, who: SRef) -> Vec<(SRef, WRef)>; // (non-transitive)
+    fn aggregates(&self, who: SRef) -> Vec<(SRef, Origin, WRef)>; // (non-transitive)
 
     /// Returns the (Dependency Graph)[DependencyGraph] of the specification
     fn graph(&self) -> &DependencyGraph;